==> step 1 crawling_links.ipynb <==
# step 1: crawling links
we took data from two websites:
1.https://www.shopbop.com/clothing/br/v=1/13266.htm
2.https://www.avenue.com/clothing/



from selenium import webdriver
from selenium.webdriver.common.keys import Keys

from webdriver_manager.chrome import ChromeDriverManager
from bs4 import BeautifulSoup
import pandas as pd
import time 
import urllib.request
import requests
from selenium.common.exceptions import NoSuchElementException
from selenium.common.exceptions import ElementClickInterceptedException

from selenium.webdriver.common.by import By

## avenue:

driverV = webdriver.Chrome(ChromeDriverManager().install())
driverV.get('https://www.avenue.com/clothing/') 

soup = BeautifulSoup(driverV.page_source, 'html.parser') 

n_pages_V= 56 #the number of pages in the website 
counter=2
prodact_list_V= []

avenue_grid = soup.find("div",{"class":"products wrapper grid products-grid"})
href_list_V= avenue_grid.find_all("div", {"class":"product-item-photo-block"})


while(counter<(n_pages_V+1)):
    for d in href_list_V:
        a=d.find("a")
        prodact_list_V.append(a['href'])
        
    driverV.get("https://www"+".avenue.com/clothing?p="+str(counter))
    time.sleep(5)
    soup = BeautifulSoup(driverV.page_source, 'html.parser')

    avenue_grid = soup.find("div",{"class":"products wrapper grid products-grid"})
    href_list_V= avenue_grid.find_all("div", {"class":"product-item-photo-block"})

    counter=counter+1

            
            

print(prodact_list_V)
#driverV.close() 


dfV = pd.DataFrame({"link V":prodact_list_V})
dfV.to_csv("avenue_link_test3.csv", index=False) 

df=pd.read_csv("avenue_link_test3.csv")
df

## shopbop:

driverS = webdriver.Chrome(ChromeDriverManager().install())
driverS.get('https://www.shopbop.com/clothing/br/v=1/13266.htm') 

soup = BeautifulSoup(driverS.page_source, 'html.parser') 

n_pages_S= 141
counter=1
prodact_list_S= []

shopbop_grid = soup.find("ol",{"class":"ProductGrid__ProductGridContainer-sc-1lp4bnm-0 cJxrHK"})
href_list_S= shopbop_grid.find_all("a")

while(counter<(n_pages_S)+1):
    for a in href_list_S:
        full_link= "https://www.shopbop.com/"+a['href']
        prodact_list_S.append(full_link)
        
    driverS.get("https://www"+".shopbop.com/clothing/br/v=1/13266.htm?offset="+str(counter*100))
    time.sleep(5)
    soup = BeautifulSoup(driverS.page_source, 'html.parser')

    shopbop_grid = soup.find("ol",{"class":"ProductGrid__ProductGridContainer-sc-1lp4bnm-0 cJxrHK"})
    href_list_S= shopbop_grid.find_all("a")

    counter=counter+1

            
            

print(prodact_list_S)
#driverS.close() 


dfS = pd.DataFrame({"link S":prodact_list_S})
dfS.to_csv("shopbop_link_test3.csv", index=False) #saving the df !!!

df=pd.read_csv("shopbop_link_test3.csv")
df

link to pages we used:

https://www.geeksforgeeks.org/
https://pandas.pydata.org/docs/index.html

https://stackoverflow.com/questions/28669482/appending-pandas-dataframes-generated-in-a-for-loop
https://www.programiz.com/python-programming/methods/string/find
https://docs.python.org/3/library/stdtypes.html

https://docs.python.org/3/library/stdtypes.html
https://www.programiz.com/python-programming/methods/string/find
https://stackoverflow.com/questions/28669482/appending-pandas-dataframes-generated-in-a-for-loop
