==> step 1.1 shopbop crawling.ipynb <==
# 1.1 shopbop crawling:
creating functiones to take: rating, size, fabric, color, name, pricre and id.

runing through all links of the website and creating a DF with the above, brand name and numerical order.

from selenium import webdriver
from selenium.webdriver.common.keys import Keys

from webdriver_manager.chrome import ChromeDriverManager
from bs4 import BeautifulSoup
import pandas as pd
import time 
import urllib.request
import requests
from selenium.common.exceptions import NoSuchElementException
from selenium.common.exceptions import ElementClickInterceptedException

from selenium.webdriver.common.by import By

driverS = webdriver.Chrome(ChromeDriverManager().install())
driverS.get('https://www.shopbop.com/clothing/br/v=1/13266.htm') 

def ratingS(soup):
    star=soup.find("meta", {"itemprop":"ratingValue"})
    if(star!= None):
        rating= star.attrs['content']
    else:
        rating=0
    return rating


def sizeS(soup):
    div=soup.find("div", {"id":"sizeList"})
    size_tags=div.find_all("div", {"class":"sizeBox"})
    if size_tags==None: return None
    size=[]
    for s in size_tags:
        size.append(s.string)
    return size

def fabricS(soup):
    div=soup.find("div", {"class":"Details__Section-sc-3us40t-0 bXftpn"})
    if div==None: return None
    fabric=div.find("span")
    if fabric==None: return None
    return fabric.string

def colorS(soup):
    div=soup.find("div", {"class":"colors"})
    color=div.find_all("img")
    if color==None: return None
    colors_list=[]
    for c in color:
        colors_list.append(c.attrs['alt'])
    return colors_list

def name_price_id(soup):
    n=p=i=None
    
    name=soup.find("div", {"id":"product-title"})
    if name!=None: 
        n=name.string
    
    price=soup.find("meta", {"itemprop":"lowPrice"})
    if price!=None: 
        p=price.attrs['content']
    
    idp=soup.find("span", {"itemprop":"sku"})
    if idp!=None: 
        i=idp.string
        
    return n, p, i


dfS=pd.read_csv("shopbop_link_test3.csv")

ratings=[]
prices=[]
product_names=[] 
brands=[] 
colors=[] 
sizes=[] 
fabrics=[] 
ids=[]

categorys=[] 

row = dict()
temp_list=[]

for num, s in enumerate(dfS["link S"]):
    driverS.get(s) 
    time.sleep(4)
    
    row = dict()
    row['num']=num
    soup = BeautifulSoup(driverS.page_source, 'html.parser')
    brands.append("shopbop")
    row['brand']= "shopbop"
    
    name, price, idp= name_price_id(soup)

    product_names.append(name)
    row['product_names']= name

    prices.append(price)
    row['prices']= price

    ids.append(idp)
    row['ids']= idp
    
    rating=ratingS(soup)
    ratings.append(rating)
    row['ratings']= rating
    
    size=sizeS(soup)
    sizes.append(size)
    row['sizes']= size
    
    fabric=fabricS(soup)
    fabrics.append(fabric)
    row['fabrics']= fabric
    
    color=colorS(soup)
    colors.append(color)
    row['colors']= color
    
    temp_list.append(row)
    
    if num%100==0:
        df = pd.DataFrame(temp_list)
        df.to_csv("shopbop_prodoct_info_temp2.csv", index=False)


df

הריצה הראשונה נתקעה אחרי 7600 הורדות, נמשיך את הריצה הבאה מאותה הנקודה:

dfS=pd.read_csv("shopbop_link_test3.csv")

ratings=[]
prices=[]
product_names=[] 
brands=[] 
colors=[] 
sizes=[] 
fabrics=[] 
ids=[]

categorys=[] 

row = dict()
temp_list=[]

for num, s in enumerate(dfS["link S"]):
    if num>7600:
        
        driverS.get(s) 
        time.sleep(4)

        row = dict()
        row['num']=num
        soup = BeautifulSoup(driverS.page_source, 'html.parser')
        brands.append("shopbop")
        row['brand']= "shopbop"

        name, price, idp= name_price_id(soup)

        product_names.append(name)
        row['product_names']= name

        prices.append(price)
        row['prices']= price

        ids.append(idp)
        row['ids']= idp

        rating=ratingS(soup)
        ratings.append(rating)
        row['ratings']= rating

        size=sizeS(soup)
        sizes.append(size)
        row['sizes']= size

        fabric=fabricS(soup)
        fabrics.append(fabric)
        row['fabrics']= fabric

        color=colorS(soup)
        colors.append(color)
        row['colors']= color

        temp_list.append(row)

        if num%100==0:
            df = pd.DataFrame(temp_list)
            df.to_csv("shopbop_prodoct_info_temp3.csv", index=False)
        

נתקע שוב אחרי 4400 הורדות(12000) 

df

dfS=pd.read_csv("shopbop_link_test3.csv")

ratings=[]
prices=[]
product_names=[] 
brands=[] 
colors=[] 
sizes=[] 
fabrics=[] 
ids=[]

categorys=[] 

row = dict()
temp_list=[]

for num, s in enumerate(dfS["link S"]):
    if num>12000:
        
        driverS.get(s) 
        time.sleep(4)

        row = dict()
        row['num']=num
        soup = BeautifulSoup(driverS.page_source, 'html.parser')
        brands.append("shopbop")
        row['brand']= "shopbop"

        name, price, idp= name_price_id(soup)

        product_names.append(name)
        row['product_names']= name

        prices.append(price)
        row['prices']= price

        ids.append(idp)
        row['ids']= idp

        rating=ratingS(soup)
        ratings.append(rating)
        row['ratings']= rating

        size=sizeS(soup)
        sizes.append(size)
        row['sizes']= size

        fabric=fabricS(soup)
        fabrics.append(fabric)
        row['fabrics']= fabric

        color=colorS(soup)
        colors.append(color)
        row['colors']= color

        temp_list.append(row)

        if num%100==0:
            df = pd.DataFrame(temp_list)
            df.to_csv("shopbop_prodoct_info_temp4.csv", index=False)
        

df

נתקע עוד פעם אחרי שהוריד 800 (12800)

dfS=pd.read_csv("shopbop_link_test3.csv")

ratings=[]
prices=[]
product_names=[] 
brands=[] 
colors=[] 
sizes=[] 
fabrics=[] 
ids=[]

categorys=[] 

row = dict()
temp_list=[]

for num, s in enumerate(dfS["link S"]):
    if num>12800:
        
        driverS.get(s) 
        time.sleep(4)

        row = dict()
        row['num']=num
        soup = BeautifulSoup(driverS.page_source, 'html.parser')
        brands.append("shopbop")
        row['brand']= "shopbop"

        name, price, idp= name_price_id(soup)

        product_names.append(name)
        row['product_names']= name

        prices.append(price)
        row['prices']= price

        ids.append(idp)
        row['ids']= idp

        rating=ratingS(soup)
        ratings.append(rating)
        row['ratings']= rating

        size=sizeS(soup)
        sizes.append(size)
        row['sizes']= size

        fabric=fabricS(soup)
        fabrics.append(fabric)
        row['fabrics']= fabric

        color=colorS(soup)
        colors.append(color)
        row['colors']= color

        temp_list.append(row)

        if num%100==0:
            df = pd.DataFrame(temp_list)
            df.to_csv("shopbop_prodoct_info_temp5.csv", index=False)
        
df

dfS=pd.read_csv("shopbop_link_test3.csv")

ratings=[]
prices=[]
product_names=[] 
brands=[] 
colors=[] 
sizes=[] 
fabrics=[] 
ids=[]

categorys=[] 

row = dict()
temp_list=[]

for num, s in enumerate(dfS["link S"]):
    if num>14100:
        
        driverS.get(s) 
        time.sleep(4)

        row = dict()
        row['num']=num
        soup = BeautifulSoup(driverS.page_source, 'html.parser')
        brands.append("shopbop")
        row['brand']= "shopbop"

        name, price, idp= name_price_id(soup)

        product_names.append(name)
        row['product_names']= name

        prices.append(price)
        row['prices']= price

        ids.append(idp)
        row['ids']= idp

        rating=ratingS(soup)
        ratings.append(rating)
        row['ratings']= rating

        size=sizeS(soup)
        sizes.append(size)
        row['sizes']= size

        fabric=fabricS(soup)
        fabrics.append(fabric)
        row['fabrics']= fabric

        color=colorS(soup)
        colors.append(color)
        row['colors']= color

        temp_list.append(row)

        if num%1==0:
            df = pd.DataFrame(temp_list)
            df.to_csv("shopbop_prodoct_info_temp6.csv", index=False)
        
df

חיברנו את כל הטבלאות לטבלה אחת:

df1= pd.read_csv("shopbop_prodoct_info_temp2.csv")
df2= pd.read_csv("shopbop_prodoct_info_temp3.csv")

df3 = df1.append(df2, ignore_index=True)

for i in range(4,7):
    df1=pd.read_csv("shopbop_prodoct_info_temp"+str(i)+".csv")
    
    df4 = df3.append(df1, ignore_index=True)
    
    df3=df4

df4

df4.to_csv("shopbop_prodoct_info_final.csv", index=False)
