==> step 1.2 avenue crawling.ipynb <==
# 1.2 shopbop crawling:
creating functiones to take: rating, size, fabric, color, name, pricre and id.
(the fabric fanction used selenium)

runing through all links of the website and creating a DF with the above, brand name and numerical order.

after finding a problom with the initial df, we redownloaded the problematic rows.

combaining the df to create a final df.

from selenium import webdriver
from selenium.webdriver.common.keys import Keys

from webdriver_manager.chrome import ChromeDriverManager
from bs4 import BeautifulSoup
import pandas as pd
import time 
import urllib.request
import requests
from selenium.common.exceptions import NoSuchElementException
from selenium.common.exceptions import ElementClickInterceptedException

from selenium.webdriver.common.by import By

import re

from selenium.webdriver.support.wait import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC

import numpy as np

driverV = webdriver.Chrome(ChromeDriverManager().install())

def ratingV(driverV):
    
    soup = BeautifulSoup(driverV.page_source, 'html.parser')
    
    div=soup.find("div", {"class":"yotpoBottomLine"})
    if div==None: return None
    full_stars= div.find_all("span", {"class":"yotpo-icon yotpo-icon-star rating-star pull-left"})
    half_stars= div.find_all("span", {"class":"yotpo-icon yotpo-icon-half-star rating-star pull-left"})
    Rating= len(full_stars)+ 0.5*len(half_stars)
    return Rating

def sizeV(soup):
    div=soup.find("ul", {"class":"control-dropdown__values-list"})
    if div==None: return 'one_size'
    size_tags=div.find_all("span", {"class":"control-dropdown__size-num"})
    if size_tags==None: return 'one_size'
    size=[]
    for s in size_tags:
        size.append(s.string)
    return size


def fabricV(driverV):
    content=None
    
    try:
        driverV.find_element(By.XPATH,'//*[@id="tab-label-content.care-title"]')
    except NoSuchElementException:
            return content
    
    soup = BeautifulSoup(driverV.page_source, 'html.parser')

    for sec in soup.find_all("script",{"type":'application/ld+json'}): # עבור תכונות ספציפיות שתגזרו

        if sec==None: return None
        text=str(sec)#convert to string
        word_to_search = '"material"'
        index = text.find(word_to_search) # find specific elemnt -> check out where are all the others
        only_details = text[index+len(word_to_search):] # without the word "matirial"
        if(index!=-1): 
            content = re.findall('"([^"]*)"', only_details)[0] # prints only the "matiral's content, part between the quets
            break
            
    return content

def colorV(soup):
    c=soup.find("div", {"class":"color color--selected"})
    if(c!= None):
        color= c.attrs['title']
    else:
        color= None
    return color

def name_price_idV(soup):
    n=p=i=None
    
    name=soup.find("span", {"class":"base"})
    if name!=None: 
        n=name.string
    
    price=soup.find("meta", {"itemprop":"price"})
    if price!=None: 
        p=price.attrs['content']
    
    idp=soup.find("span", {"class":"sku"})
    if idp!=None: 
        i=idp.string
        
    return n, p, i

dead_page_list=[2555,2558,2418,2397,2398,2278,2214,2215,2216,2158,1468,638,639,260]

first try:

dfV=pd.read_csv("avenue_link_test3.csv")

ratings=[]
prices=[]
product_names=[] 
brands=[] 
colors=[] 
sizes=[] 
fabrics=[] 
ids=[]

categorys=[] 

row = dict()
temp_list=[]

for num, v in enumerate(dfV["link V"]):
    driverV.get(v) 
    time.sleep(4)
    
    row = dict()
    row['num']=num
    soup = BeautifulSoup(driverV.page_source, 'html.parser')
    brands.append("avenue")
    row['brand']= "avenue"
    
    name, price, idp= name_price_idV(soup)

    product_names.append(name)
    row['product_names']= name

    prices.append(price)
    row['prices']= price

    ids.append(idp)
    row['ids']= idp
    
    rating=ratingV(soup)
    ratings.append(rating)
    row['ratings']= rating
    
    size=sizeV(soup)
    sizes.append(size)
    row['sizes']= size
    
    try:
        fabric=fabricV(driverV)
    exepct
    fabrics.append(fabric)
    row['fabrics']= fabric
    
    color=colorV(soup)
    colors.append(color)
    row['colors']= color
    
    temp_list.append(row)
    
    if num%100==0:
        df = pd.DataFrame(temp_list)
        df.to_csv("avenue_prodoct_info_final.csv", index=False)

df=pd.read_csv("avenue_prodoct_info_final.csv")
df

df.shape

there was a problom with some of the rows downloaded, especially at the rating column, therefore we redawnloaded those rows.

df_nan1=df.dropna()
df1= df_nan1[(df_nan1.ratings== 0.0)]

num_list=[]
for row in df1.num:
    num_list.append(row)
# num_list

dfV=pd.read_csv("avenue_link_test3.csv")

ratings=[]
prices=[]
product_names=[] 
brands=[] 
colors=[] 
sizes=[] 
fabrics=[] 
ids=[]

row = dict()
temp_list=[]
errors_links = []

for num, v in enumerate(dfV["link V"]):
    
    if num in num_list:
        
        
        driverV.get(v) 
        time.sleep(5)
        soup = BeautifulSoup(driverV.page_source, 'html.parser')
        row = dict()
        row['num']=num

        brands.append("avenue")
        row['brand']= "avenue"

        name, price, idp= name_price_idV(soup)

        product_names.append(name)
        row['product_names']= name

        prices.append(price)
        row['prices']= price

        ids.append(idp)
        row['ids']= idp

        rating=ratingV(driverV)
        ratings.append(rating)
        row['ratings']= rating

        size=sizeV(soup)
        sizes.append(size)
        row['sizes']= size

        if (size==None):
            
            errors_links.append(num)
            print(num)


        fabric=fabricV(driverV)
        fabrics.append(fabric)
        row['fabrics']= fabric

        color=colorV(soup)
        colors.append(color)
        row['colors']= color

        temp_list.append(row)

        if num%10==0:
            df = pd.DataFrame(temp_list)
            df.to_csv("avenue_prodoct_info_temp_fix20.csv", index=False)

        
df

dfV=pd.read_csv("avenue_link_test3.csv")

ratings=[]
prices=[]
product_names=[] 
brands=[] 
colors=[] 
sizes=[] 
fabrics=[] 
ids=[]

categorys=[] 

row = dict()
temp_list=[]
errors_links = []

for num, v in enumerate(dfV["link V"]):
    
    if num>2630:
    
        if num in num_list:


            driverV.get(v) 
            time.sleep(5)
            soup = BeautifulSoup(driverV.page_source, 'html.parser')
            row = dict()
            row['num']=num

            brands.append("avenue")
            row['brand']= "avenue"

            name, price, idp= name_price_idV(soup)

            product_names.append(name)
            row['product_names']= name

            prices.append(price)
            row['prices']= price

            ids.append(idp)
            row['ids']= idp

            rating=ratingV(driverV)
            ratings.append(rating)
            row['ratings']= rating

            size=sizeV(soup)
            sizes.append(size)
            row['sizes']= size

            if (size==None):

                errors_links.append(num)
                print(num)


            fabric=fabricV(driverV)
            fabrics.append(fabric)
            row['fabrics']= fabric

            color=colorV(soup)
            colors.append(color)
            row['colors']= color

            temp_list.append(row)

            if num%1==0:
                df = pd.DataFrame(temp_list)
                df.to_csv("avenue_prodoct_info_temp_fix25.csv", index=False)

        
df

running again over the ranked ones to complete the table (all of the ones we didnt run over at the previouse run)

dfV=pd.read_csv("avenue_link_test3.csv")
count=0
for num, v in enumerate(dfV["link V"]):
    
    if num not in dead_page_list and num not in num_list:
        count=count+1
        
count

dfV=pd.read_csv("avenue_link_test3.csv")

ratings=[]
prices=[]
product_names=[] 
brands=[] 
colors=[] 
sizes=[] 
fabrics=[] 
ids=[]

categorys=[] 

row = dict()
temp_list=[]
errors_links = []

for num, v in enumerate(dfV["link V"]):
    
    if num not in dead_page_list and num not in num_list:

        driverV.get(v) 
        time.sleep(5)
        soup = BeautifulSoup(driverV.page_source, 'html.parser')
        row = dict()
        row['num']=num

        brands.append("avenue")
        row['brand']= "avenue"

        name, price, idp= name_price_idV(soup)

        product_names.append(name)
        row['product_names']= name

        prices.append(price)
        row['prices']= price

        ids.append(idp)
        row['ids']= idp

        rating=ratingV(driverV)
        ratings.append(rating)
        row['ratings']= rating

        size=sizeV(soup)
        sizes.append(size)
        row['sizes']= size

        if (size==None):

            errors_links.append(num)
            print(num)


        fabric=fabricV(driverV)
        fabrics.append(fabric)
        row['fabrics']= fabric

        color=colorV(soup)
        colors.append(color)
        row['colors']= color

        temp_list.append(row)

        if num%1==0:
            df = pd.DataFrame(temp_list)
            df.to_csv("avenue_prodoct_info_temp_fix30.csv", index=False)


df

df=pd.read_csv("avenue_prodoct_info_temp_fix30.csv")
df

dfV=pd.read_csv("avenue_link_test3.csv")

ratings=[]
prices=[]
product_names=[] 
brands=[] 
colors=[] 
sizes=[] 
fabrics=[] 
ids=[]

categorys=[] 

row = dict()
temp_list=[]
errors_links = []

for num, v in enumerate(dfV["link V"]):
    
    if num not in dead_page_list and num not in num_list and num>357:

        driverV.get(v) 
        time.sleep(5)
        soup = BeautifulSoup(driverV.page_source, 'html.parser')
        row = dict()
        row['num']=num

        brands.append("avenue")
        row['brand']= "avenue"

        name, price, idp= name_price_idV(soup)

        product_names.append(name)
        row['product_names']= name

        prices.append(price)
        row['prices']= price

        ids.append(idp)
        row['ids']= idp

        rating=ratingV(driverV)
        ratings.append(rating)
        row['ratings']= rating

        size=sizeV(soup)
        sizes.append(size)
        row['sizes']= size

        if (size==None):

            errors_links.append(num)
            print(num)


        fabric=fabricV(driverV)
        fabrics.append(fabric)
        row['fabrics']= fabric

        color=colorV(soup)
        colors.append(color)
        row['colors']= color

        temp_list.append(row)

        if num%1==0:
            df = pd.DataFrame(temp_list)
            df.to_csv("avenue_prodoct_info_temp_fix35.csv", index=False)


df

combained final table:

df1= pd.read_csv("avenue_prodoct_info_temp_fix20.csv")
df2= pd.read_csv("avenue_prodoct_info_temp_fix25.csv")
df3= pd.read_csv("avenue_prodoct_info_temp_fix30.csv")
df4= pd.read_csv("avenue_prodoct_info_temp_fix35.csv")

df_final = df1.append([df2, df3,df4], ignore_index=True)

df_final

df_final.to_csv("avenue_prodoct_info_final_fixed_table.csv", index=False)
