==> step 2.1 shopbop cleaning.ipynb <==
# step 2.1 shopbop cleaning # 

cleaning, spliting lists columns, deleting and adding columns to create the final df for the shopbop website.

import pandas as pd
import re

df= pd.read_csv("shopbop_prodoct_info_final.csv")
df

## Nan and duplicates values

טיפול בנתונים החסרים בטבלה ובנתונים הכפולים.
במידה של כפילות הורדה של שורות כפולות, ובמידה של נתון חסר טיפול בהתאם לצורך

df.info()

ניתן לראות שכל החסרים נמצאים באותה עמודה,
עמודת הבד שאינה ניתנת לחיזוי כיוון שתרכובת הבד שונה לחלוטין בכל פריט ביגוד והכנסה של ערך שכיח\ מילוי ע"פ עוקב\קודם עלולה ליצור הטייה משמעותית בנכונות ההתפלגות
לכן בחרנו להוריד שורות אלו.

df_nan1=df.dropna()
df_nan1.shape

df_nan1.to_csv("shopbop_cleaning_temp1.csv", index=False)

df1= df[(df.ratings!= 0.0)]
# df1
df1.shape

df2= df_nan1[(df_nan1.ratings!= 0.0)]
df2.shape

לאחר בדיקת התפלגות הפריטים המדורגים והלא מדורגים לפני ואחרי ההורדה גילינו שאחוז ההתפלגות נשמר גם לאחר ההורדה לכן הסקנו שההורדה אינה פוגעת בנכונות הנתונים.

df_nan1.duplicated().sum()

df_nan1

## Regex cleaning

ניקוי "רעשים" בעמודות בעזרת ספריית re


df_nan1.replace({'fabrics': r'Fabric: ?'}, {'fabrics': ''}, regex=True, inplace=True)

df_nan1

df_nan1.replace({'sizes': r"\\n                        "}, {'sizes': ''}, regex=True, inplace=True)

df_nan1

## deleting and adding columns 
הורדת שורות לא רלוונטיות כמו שורת המספר השימשה לצורך סידורי בשלב הקודם והוספת עמודות חדשות היכולות לעזור בשלב המודל כמו עמודת הקטגוריה.

df_nan1.drop("num", inplace=True, axis=1)

df_nan1

Dresses=['Dress','Chemise']
Jackets_Coats=['Hoodie','Jacket', 'Coat','Blazer','Cardigan']
Jumpsuits_Rompers=['Jumpsuit', 'Romper','Set']
Lingerie=['Lingerie', 'Bra', 'Thong','Briefs','Bodysuit', 'Cami','Bralette','Boyshorts','G-String','Tights','Panties']
Maternity=['Maternity']
Pants=['Leggings', 'Skort', 'Pants', 'Biker Shorts', 'Shorts','Sweatpants', 'Joggers','LEGGING','Jeans', 'Pant']
Skirts=['Skirts','Skirt']
Sleepwear=['Sleepwear','Pajama Set','PJ Set','Mask','Slip','Tunic']
Sweaters_Knits=['Sweater','Knit']
Swimsuits_Cover_Ups=['Swimsuit','Bikini','Bottoms']
Tops=['Sweatshirt', 'Shirt', 'Tee', 'Top', 'Muscle Tank', 'Tank', 'Crop Top','Pullover','Crew','Turtleneck','Blouse','TEE' ]

category=[]
flag=0

category_names={'Dresses':Dresses,'Jackets_Coats':Jackets_Coats,'Lingerie':Lingerie, 'Jumpsuits_Rompers':Jumpsuits_Rompers,
                'Maternity':Maternity,'Pants':Pants,'Skirts':Skirts,'Sleepwear':Sleepwear,
                'Sweaters_Knits':Sweaters_Knits,'Swimsuits_Cover_Ups':Swimsuits_Cover_Ups,'Tops':Tops}

for name in df_nan1.product_names:  
    for n in category_names:
        for item in category_names[n]:
            m=re.search(item,name)
            if m!=None:
                flag=1
                category.append(n)
                break
        if flag==1:
            break
    if flag==0:
        category.append(None)
    else: flag=0
                      

df_nan1['category'] = category

df_nan1

לאחר יצירת עמודת הקטגוריה מורידים את העמודות ללא קטגוריה, גם כאן מאותן הסיבות החלטנו להוריד ולא למלא חסרים,
ובנוסף לאחר בדיקה ידנית גילינו שרוב הפריטים שלא דורגו היו פריטים שאכן אינם שיכיים לקטגוריות ביגוד ולא היו רלוונטים מלכתחילה. 

df_nan2=df_nan1.dropna()
df_nan2.shape

df5=df_nan2[(df_nan2.ratings!= 0.0)]
df5.shape

df_nan2.to_csv("shopbop_cleaning_final_clean1", index=False)

## Data exploration

df=pd.read_csv("shopbop_cleaning_final_clean1")
df

df.describe()

df.describe(include=['O'])

## Binary split of categorical columns
עמודות השמורות כרשימה של ערכים קטגוריאלים יומרו למספר עמודות בינאריות לפי ערכים.

XXS_list=[0]*8855
XS_list=[0]*8855
P_list=[0]*8855
S_list=[0]*8855
M_list=[0]*8855
L_list=[0]*8855
XL_list=[0]*8855
XXL_list=[0]*8855
XXXL_list=[0]*8855
XXXXL_list=[0]*8855
X5L_list=[0]*8855
X6L_list=[0]*8855
onesize_list=[0]*8855


for num, row in enumerate(df.sizes):
    for s in range(len(row)):
        if row[s]=='00' or row[s]=='30' or row[s]=='34' or row[s]=='XXS':
            XXS_list[num]=1
            
        if row[s]=='0' or row[s]=='2' or row[s]=='32' or row[s]=='36' or row[s]=='38' or row[s]=='XS':
            XS_list[num]=1
            
        if row[s]=='P':
            P_list[num]=1
            
        if row[s]=='4' or row[s]=='6' or row[s]=='40' or row[s]=='42' or row[s]=='S':
            S_list[num]=1
            
        if row[s]=='8' or row[s]=='10' or row[s]=='44' or row[s]=='46' or row[s]=='M':
            M_list[num]=1
            
        if row[s]=='12' or row[s]=='48' or row[s]=='L':
            L_list[num]=1
            
        if row[s]=='14' or row[s]=='50' or row[s]=='XL':
            XL_list[num]=1
            
        if row[s]=='16' or row[s]=='52' or row[s]=='XXL':
            XXL_list[num]=1
            
        if row[s]=='18' or row[s]=='20' or row[s]=='54' or row[s]=='56' or row[s]=='XXXL':
            XXXL_list[num]=1
            
        if row[s]=='22' or row[s]=='24' or row[s]=='26' or row[s]=='28' or row[s]=='58' or row[s]=='60' or row[s]=='XXXXL':
            XXXXL_list[num]=1           
            
df_size=pd.DataFrame({"XXS":XXS_list, "XS":XS_list, "S":S_list, "P":P_list, "M":M_list, "L":L_list, "XL":XL_list,
                      "XXL":XXL_list, "XXXL":XXXL_list, "XXXXL":XXXXL_list, "One_size":onesize_list,"5XL":X5L_list,"6XL":X6L_list})      

df_size

df_size.to_csv("shopbop_cleaning_final_size_table.csv", index=False)

BLUE_list=[0]*8855
BLACK_list=[0]*8855
WHITE_list=[0]*8855
GREY_list=[0]*8855
RED_list=[0]*8855
BROWN_list=[0]*8855
ORANGE_list=[0]*8855
PINK_list=[0]*8855
GREEN_list=[0]*8855
PURPLE_list=[0]*8855
YELLOW_list=[0]*8855

OTHER_COLORS_list=[0]*8855


for num, row in enumerate(df.colors):
    
    flag=0

    if re.search('Blue|Navy|Midnight|Cobalt|Giungla|Denim|Light Lagoon|Caribe',row):
        BLUE_list[num]=1
        flag=1

    if re.search('Black|Carbon|Noir|Coal|Licorice|Jet',row):
        BLACK_list[num]=1
        flag=1

    if re.search('Grey|Charcoal|Smoke|Veer|Iron',row):
        GREY_list[num]=1
        flag=1

    if re.search('White|Ivory|Sugar|Blanc|Nylon|Champagne|Cream|Clear|Bone|Porcelain|Fair|Eggshell',row):
        WHITE_list[num]=1
        flag=1

    if re.search('Brown|Caramel|Saddle|Coco|Creme|Almond|Frappe|Cafe|Taupe|Tan|Cashmere|Camel|Mocha|Toffe|Cinnamon|Copper|Chai|Latte|Gobi|Mandorla|Beige',row):
        BROWN_list[num]=1
        flag=1

    if re.search('Orange|Honey|Peach',row):
        ORANGE_list[num]=1
        flag=1

    if re.search('Pink|Bare|Tuberose|Rose|Blush|Petal',row):
        PINK_list[num]=1
        flag=1
        
    if re.search('Green|Khaki|Army|Cactus|Sage',row):
        GREEN_list[num]=1
        flag=1
        
    if re.search('Purple|Indigo|Lilac',row):
        PURPLE_list[num]=1
        flag=1
        
    if re.search('Red',row):
        RED_list[num]=1
        flag=1
        
    if re.search('Yellow|Buttercream',row):
        YELLOW_list[num]=1
        flag=1
        
    if flag==0:
        OTHER_COLORS_list[num]=1


            
df_color=pd.DataFrame({"Blue":BLUE_list, 'Black':BLACK_list, "Grey":GREY_list, 'White':WHITE_list, 
                      'Brown':BROWN_list, 'Orange':ORANGE_list, 'Pink':PINK_list, 'Green':GREEN_list,
                     'Purple':PURPLE_list, 'Red':RED_list, 'Yellow':YELLOW_list, 'other_colors':OTHER_COLORS_list})      

df_color

df_color.to_csv("shopbop_cleaning_final_color_table.csv", index=False)

## Categorization of the y label
קטגוריזציה של ערך רציף דירוג ל6 קטגוריות מרכזיות

labeled_rating=[]

for num, row in enumerate(df.ratings):
    
    if row<=0.5:
        labeled_rating.append(0)
        
    if row>0.5 and row<=1.5:
        labeled_rating.append(1)
        
    if row>1.5 and row<=2.5:
        labeled_rating.append(2)
        
    if row>2.5 and row<=3.5:
        labeled_rating.append(3)
    
    if row>3.5 and row<=4.5:
        labeled_rating.append(4)
        
    if row>4.5:
        labeled_rating.append(5)
        
len(labeled_rating)

df["ratings"]=labeled_rating

## Final table

חיבור כל הטבלאות השונות וניקוי של הטבלאות שלא בשימוש יותר

df_size= pd.read_csv("shopbop_cleaning_final_size_table.csv")
df_color=pd.read_csv("shopbop_cleaning_final_color_table.csv")

base_df=pd.read_csv("shopbop_cleaning_final_clean1")
copy_base_df=base_df.copy()

copy_base_df.to_csv("shopbop_cleaning_2_table.csv", index=False)

copy_base_df.drop(["sizes", "colors"], inplace=True, axis=1)

copy_base_df["ratings"]=labeled_rating

df = pd.concat([copy_base_df, df_size, df_color], axis=1)

df.to_csv("shopbop_cleaning_final_final_table.csv", index=False)

df
