==> step 2.2 avenue cleaning.ipynb <==
# step 2.2 avenue cleaning
cleaning, spliting lists columns, deleting and adding columns to create the final df for the avenue website.

import pandas as pd
import re

# Nan and duplicates values
טיפול בנתונים החסרים בטבלה ובנתונים הכפולים. במידה של כפילות הורדה של שורות כפולות, ובמידה של נתון חסר טיפול בהתאם לצורך

df= pd.read_csv("avenue_prodoct_info_final_fixed_table.csv")
df.shape

df.info()

df[(df.ratings!= 0.0)].shape

ניתן לראות שהחוסרים נמצאים בעמודות: שם הפריט, מחיר, מספר זיהוי, דירוג, בד וצבע.

בעמודת הדירוג שהיא גם עמודת המטרה שלנו לא ניתן להכניס שרירותית ערכים בכדי למלא את החסרים כיוון שהדבר עלול לגרום להטייה משמעותית בנכונות המודל.

בעמודות השם ומס הזיהוי לא ניתן להשלים חוסרים כיוון שלכל פריט קיים ערך אינדיוידואלי.

בעמודת הבד בחרנו שלא להשלים חוסרים כיוון שלא ניתן לחזות את תרכובת הבד השונה לחלוטין בכל פריט ביגוד והכנסה של ערך שכיח\ מילוי ע"פ עוקב\קודם עלולה לגרום להטייה.

עמודת המחיר: מילוי החוסרים ע"י עוקב\קודם בטבלה אינו רלוונטי כיוון שסדר הפריטים באתר אינו ע"פ מחיר וכך גם בטבלה שהורכשה מהאתר, בנוסף השלמה על ידי שכיח העמודה גם אינו רלוונטי, נשארנו רק עם מילוי ערכים חסרים, מילוי ע"י חציון\ממוצע כל הערכים בטבלה אינו נכון כיוון שהפריטים מחולקים לפי קטגוריה ולכל קטגוריה טווח מחירים שונה, לכן הוחלט להכניס ערך זמני * עד לטיפול מאוחר יותר בנתונים אלו על סמך עמודת הקטגוריה.

עמודת הצבע: פריטים ללא צבע הינם בדרך כלל פריטים אשר מוצגים כהדפס או דוגמא מיוחדת שלא מצריכה פירוט צבע ולכן עמודות אלו ימולאו בסימן * אשר מסמל על קטגורית צבע אחר.

df.prices.fillna('*', inplace=True)
df.colors.fillna('*', inplace=True)

df_nan1=df.dropna()
df_nan1.shape

df_nan1[(df_nan1.ratings!= 0.0)].shape

df_nan1.to_csv("avenue_cleaning_temp1.csv", index=False)

df_nan1.duplicated().sum()

# Regex cleaning
ניקוי "רעשים" בעמודות בעזרת ספריית re

df_nan1.replace({'ids': r'Style # '}, {'ids': ''}, regex=True, inplace=True)

# deleting and adding columns
הורדת שורות לא רלוונטיות כמו שורת המספר השימשה לצורך סידורי בשלב הקודם והוספת עמודות חדשות היכולות לעזור בשלב המודל כמו עמודת הקטגוריה.

df_nan1.drop("num", inplace=True, axis=1)

Dresses=['Dress','Chemise','Maxi']
Jackets_Coats=['Hoodie','Jacket', 'Coat','Blazer','Cardigan','Cape','Poncho','Peacoat','Trench','Parka','Capri']
Jumpsuits_Rompers=['Jumpsuit', 'Romper','Set',' Jumpsuit -','Tux','Piece']
Lingerie=['Lingerie', 'Bra', 'Thong','Briefs','Bodysuit', 'Cami','Bralette','Boyshorts','G-String','Tights','Panties']
Maternity=['Maternity']
Pants=['Leggings','Legging','Trouser','Culotte', 'Skort', 'Pants', 'Biker Shorts','Short', 'Shorts','Sweatpants','Jogger' ,'Joggers','LEGGING','Jeans', 'Pant','Jegging','Jean']
Skirts=['Skirts','Skirt','Skirt-']
Sleepwear=['Sleepwear','Pajama Set','PJ Set','Mask','Slip','Tunic']
Sweaters_Knits=['Sweater','Knit']
Swimsuits_Cover_Ups=['Swimsuit','Bikini','Bottoms','Swim']
Tops=['Sweatshirt', 'Shirt', 'Tee', 'Top', 'Muscle Tank', 'Tank', 'Crop Top','Pullover','Crew','Turtleneck','Blouse','TEE','Vest' ]

category=[]
flag=0

category_names={'Dresses':Dresses,'Jackets_Coats':Jackets_Coats,'Lingerie':Lingerie,'Jumpsuits_Rompers':Jumpsuits_Rompers,
                'Maternity':Maternity,'Pants':Pants,'Skirts':Skirts,'Sleepwear':Sleepwear,
                'Sweaters_Knits':Sweaters_Knits,'Swimsuits_Cover_Ups':Swimsuits_Cover_Ups,'Tops':Tops}

for name in df_nan1.product_names:  
    for n in category_names:
        for item in category_names[n]:
            m=re.search(item,name)
            if m!=None:
                flag=1
                category.append(n)
                break
        if flag==1:
            break
    if flag==0:
        category.append(None)
    else: flag=0
            

df_nan1['category'] = category

df_nan2=df_nan1.dropna()
df_nan2.shape

לאחר יצירת עמודת הקטגוריה מורידים את העמודות ללא קטגוריה, גם כאן מאותן הסיבות החלטנו להוריד ולא למלא חסרים, ובנוסף לאחר בדיקה ידנית גילינו שרוב הפריטים שלא דורגו היו פריטים שאכן אינם שיכיים לקטגוריות ביגוד ולא היו רלוונטים מלכתחילה.
בנוסף רק 13 פריטים דורגו ללא קטגוריה דבר אשר לא פוגע בהתפלגות הנתונים בטבלה.

## Price column based category column
סידור הערכים החסרים בעמודת המחיר ע"פ הממוצע של כל קטגורית ביגוד. 

category_names=['Dresses','Jackets_Coats','Lingerie','Jumpsuits_Rompers',
                'Maternity','Pants','Skirts','Sleepwear',
                'Sweaters_Knits','Swimsuits_Cover_Ups','Tops']

category_dict={'Dresses':0, 'Jackets_Coats':1, 'Lingerie':2, 'Jumpsuits_Rompers':3, 'Maternity':4, 'Pants':5,
              'Skirts':6, 'Sleepwear':7, 'Sweaters_Knits':8, 'Swimsuits_Cover_Ups':9, 'Tops':10}

sum_list=[0]*11
num_list=[0]*11
mean_list=[0]*11

category_df=pd.DataFrame({'name':category_names, 'num':num_list ,'sum':sum_list, 'mean': mean_list})

for ind in df_nan2.index:
    if(df_nan2["prices"][ind]!='*'):
        #num per category
        category_df["num"][category_dict[df_nan2["category"][ind]]]= category_df["num"][category_dict[df_nan2["category"][ind]]]+1
        #sum per category
        category_df["sum"][category_dict[df_nan2["category"][ind]]]= category_df["sum"][category_dict[df_nan2["category"][ind]]]+df_nan2["prices"][ind]

for ind in range(0,11):
    category_df["mean"][ind]=category_df["sum"][ind]/category_df["num"][ind]
    

category_df  

for ind in df_nan2.index:
    if(df_nan2["prices"][ind]=='*'):
        df_nan2["prices"][ind ]= category_df["mean"][category_dict[df_nan2["category"][ind]]]

df_nan2[(df_nan2.prices== "*")].shape

# Categorization of the y label
קטגוריזציה של ערך רציף דירוג ל6 קטגוריות מרכזיות

labeled_rating=[]

for num, row in enumerate(df_nan2.ratings):
    
    if row<=0.5:
        labeled_rating.append(0)
        
    if row>0.5 and row<=1.5:
        labeled_rating.append(1)
        
    if row>1.5 and row<=2.5:
        labeled_rating.append(2)
        
    if row>2.5 and row<=3.5:
        labeled_rating.append(3)
    
    if row>3.5 and row<=4.5:
        labeled_rating.append(4)
        
    if row>4.5:
        labeled_rating.append(5)
        
len(labeled_rating)

df_nan2["ratings"]=labeled_rating

df_nan2.to_csv("avenue_cleaning_final_clean1",index=False)

df_nan2

# Binary split of categorical columns
עמודות השמורות כרשימה של ערכים קטגוריאלים יומרו למספר עמודות בינאריות לפי ערכים.

XXS_list=[0]*2594
XS_list=[0]*2594
P_list=[0]*2594
S_list=[0]*2594
M_list=[0]*2594
L_list=[0]*2594
XL_list=[0]*2594
XXL_list=[0]*2594
XXXL_list=[0]*2594
XXXXL_list=[0]*2594
X5L_list=[0]*2594
X6L_list=[0]*2594
onesize_list=[0]*2594


for num, row in enumerate(df_nan2.sizes):
    for s in range(len(row)):
        
        if re.search('One Size|\[None\]',row):
            onesize_list[num]=1

        if re.search('14|16|14-16|0X',row):
            L_list[num]=1

        if re.search('18|20|18-20|1X',row):
            XL_list[num]=1

        if re.search('22|24|22-24|2X',row):
            XXL_list[num]=1
 
        if re.search('26|28|26-28|3X',row):
            XXXL_list[num]=1

        if re.search('30|32|30-32|4X',row):
            XXXXL_list[num]=1

        if re.search('34|36|34-36|5X',row):
            X5L_list[num]=1

        if re.search('38|40|38-40|6X',row):
            X6L_list[num]=1
            
df_size=pd.DataFrame({"XXS":XXS_list, "XS":XS_list, "S":S_list, "P":P_list, "M":M_list, "L":L_list, "XL":XL_list,
                      "XXL":XXL_list, "XXXL":XXXL_list, "XXXXL":XXXXL_list, "One_size":onesize_list,"5XL":X5L_list,"6XL":X6L_list})      

df_size

df_size.to_csv("avenue_cleaning_final_size_table.csv", index=False)

BLUE_list=[0]*2594
BLACK_list=[0]*2594
WHITE_list=[0]*2594
GREY_list=[0]*2594
RED_list=[0]*2594
BROWN_list=[0]*2594
ORANGE_list=[0]*2594
PINK_list=[0]*2594
GREEN_list=[0]*2594
PURPLE_list=[0]*2594
YELLOW_list=[0]*2594

OTHER_COLORS_list=[0]*2594


for num, row in enumerate(df_nan2.colors):
    
    flag=0

    if re.search('Blue|Blues|Ocean|Dark Sea|Aqua|Quiet Night|Sapphire|Navy Bouquet|Sky|Av Nvy|Light Denim|Mid Wash|Navy|Midnight|Cobalt|Giungla|Denim|Light Lagoon|Caribe|Dark Indigo|Indigo',row):
        BLUE_list[num]=1
        flag=1

    if re.search('Black|Av Blk|Darkwash|Dark|Carbon|Noir|Coal|Licorice|Jet|After Dark|Ink',row):
        BLACK_list[num]=1
        flag=1

    if re.search('Grey|Steel Blk|Silver|Charcoal|Smoke|Veer|Iron|Charcoal Marle|Chain Reaction|Oatmeal|Steel|Slate|Smoke',row):
        GREY_list[num]=1
        flag=1

    if re.search('White|Ivory|Sugar|Blanc|Nylon|Champagne|Cream|Clear|Bone|Porcelain|Fair|Eggshell',row):
        WHITE_list[num]=1
        flag=1

    if re.search('Terracotta|Chocolate|Sands Of Time|Brown|Caramel|Saddle|Coco|Creme|Almond|Frappe|Cafe|Taupe|Tan|Cashmere|Camel|Mocha|Toffe|Cinnamon|Copper|Chai|Latte|Gobi|Mandorla|Beige',row):
        BROWN_list[num]=1
        flag=1

    if re.search('Orange|Honey|Peach|Ginger|Peachy|Coral',row):
        ORANGE_list[num]=1
        flag=1

    if re.search('Pink|Bare|Tuberose|Rose|Blush|Petal|Luna Floral|Mulberry|Rouge|Magenta|Petunia',row):
        PINK_list[num]=1
        flag=1
        
    if re.search('Green|Khaki|Army|Mint|Cactus|Sage|Teal|Emerald Floral|Yucca Palm|Emerald|Jade|Pine|Olive',row):
        GREEN_list[num]=1
        flag=1
        
    if re.search('Purple|Indigo|Lilac|Sangria|Grape|Iris|Plum|Eggplant Blk|Misty Plum|Malaga|Amethyst|Purp|Fuschia Red',row):
        PURPLE_list[num]=1
        flag=1
        
    if re.search('Red|Berry|Burgundy|Bordeaux|Ruby|ruby|Scarlett',row):
        RED_list[num]=1
        flag=1
        
    if re.search('Yellow|Buttercream|Gold|Saffron|Amber|Buff',row):
        YELLOW_list[num]=1
        flag=1
        
    if flag==0:
        OTHER_COLORS_list[num]=1


            
df_color=pd.DataFrame({"Blue":BLUE_list, 'Black':BLACK_list, "Grey":GREY_list, 'White':WHITE_list, 
                      'Brown':BROWN_list, 'Orange':ORANGE_list, 'Pink':PINK_list, 'Green':GREEN_list,
                     'Purple':PURPLE_list, 'Red':RED_list, 'Yellow':YELLOW_list, 'other_colors':OTHER_COLORS_list})      

df_color

df_color.to_csv("avenue_cleaning_final_color_table.csv", index=False)

# Final table
חיבור כל הטבלאות השונות וניקוי של הטבלאות שלא בשימוש יותר

df_size= pd.read_csv("avenue_cleaning_final_size_table.csv")
df_color=pd.read_csv("avenue_cleaning_final_color_table.csv")

base_df=pd.read_csv("avenue_cleaning_final_clean1")
copy_base_df=base_df.copy()

copy_base_df.to_csv("avenue_cleaning_1_table.csv", index=False)

copy_base_df.drop(["sizes", "colors"], inplace=True, axis=1)

df = pd.concat([copy_base_df, df_size, df_color], axis=1)

df.to_csv("avenue_cleaning_final_table.csv", index=False)


df
