==> step 2.3 combining dfs.ipynb <==
# step 2.3 combining dfs
combining the two df of the shopbop, avenue websites.

after combining, cleaning, and organizing the data, columns that needed to be taken care of after the combination.

import pandas as pd

from sklearn.feature_extraction.text import CountVectorizer
from sklearn import preprocessing

#for the stop words
from sklearn.feature_extraction import text 

## combaining dfs

dfS=pd.read_csv("shopbop_cleaning_final_final_table.csv")
dfV=pd.read_csv("avenue_cleaning_final_table.csv")

df = dfS.append(dfV, ignore_index=True)
df.to_csv("finale_df.csv")

df 

## fabric column
וקטוריזציה של עמודת הבד על ידי מציאת 20 הסוגים הכי שכיחים ויצוג בינארי שלהם  

df.replace({'fabrics': r'-'}, {'fabrics': ''}, regex=True, inplace=True)

my_additional_stop_words=["100", "95", "main",'fabric','outer','97','trim','slinky','ribbed','acrylic']

stop_words = text.ENGLISH_STOP_WORDS.union(my_additional_stop_words)

vectorizer= CountVectorizer(lowercase=True, stop_words=stop_words , min_df=0.01, max_features=20,binary=True)

x= vectorizer.fit_transform(df['fabrics'])

x.toarray()

vectorizer.get_feature_names()

df_fabric = pd.DataFrame(x.toarray())
df_fabric.columns = vectorizer.get_feature_names()
df_fabric

df_fabric.to_csv("final_fabric_table.csv", index=False)

dataset = pd.concat([df,df_fabric], axis=1)
dataset

dataset.to_csv("final_df_with_fabric_vectors.csv", index=False)
