==> step 4.1 Final Cleaning.ipynb <==
# step 4.1 Final Cleaning
prepering the df for the model.

import pandas as pd

df = pd.read_csv("final_df_with_fabric_vectors.csv")

## Categorical columns

טיפול בעמודות קטגוריאליות, קטגוריזציה של עמודות קטגוריאליות כמו עמודת המותג, והורדת עמודות שלא ניתן לקטרג כמו עמודת השם, מספר הזיהוי ותרכובת הבד 

brand_normalized=[]

for row in df.brand:
    if row=="shopbop":
        brand_normalized.append(1)
        
    if row=="avenue":
        brand_normalized.append(2)
        
df.brand=brand_normalized  
df

category_dict={1:'Dresses', 2:'Jackets_Coats', 3:'Jumpsuits_Rompers', 4:'Lingerie', 5:'Maternity', 6:'Pants',
              7:'Skirts', 8:'Sleepwear', 9:'Sweaters_Knits', 10:'Swimsuits_Cover_Ups', 11:'Tops'}
normalized_categorys=[]

for row in df.category:
    if row=='Dresses': normalized_categorys.append(1)
    if row=='Jackets_Coats': normalized_categorys.append(2)    
    if row=='Jumpsuits_Rompers': normalized_categorys.append(3)
    if row=='Lingerie': normalized_categorys.append(4)    
    if row=='Maternity': normalized_categorys.append(5)    
    if row=='Pants': normalized_categorys.append(6)    
    if row=='Skirts': normalized_categorys.append(7)  
    if row=='Sleepwear': normalized_categorys.append(8)   
    if row=='Sweaters_Knits': normalized_categorys.append(9)    
    if row=='Swimsuits_Cover_Ups': normalized_categorys.append(10)    
    if row=='Tops': normalized_categorys.append(11)
        
df["category"]=normalized_categorys

df_non_cat=df.copy()
df_non_cat.to_csv("dataset_for_model.csv", index=False)
df_non_cat.drop(["fabrics","product_names","ids"], inplace=True, axis=1)
df_non_cat

## Continuous variable
טיפול במשתנים רציפים כמו עמודת המחיר, נירמול של העמודה. 

# df on input should contain only one column with the price data (plus dataframe index)
# return: y will be a column in a dataframe
def normalize_col(df):
    min = df.min()
    max = df.max()
    x = df 
    y = (x - min) / (max - min)
    
    return y

df_non_cat['norm_prices']= normalize_col(df_non_cat.prices)
df_non_cat['norm_prices']

df_non_cat.drop(["prices"], inplace=True, axis=1)
df_non_cat

df_non_cat.to_csv("finale_normalized_df.csv", index=False)
