==> step 4.2 Machine Learning - Model.ipynb <==
# step 4.2 Machine Learning - Model
creating a model and optimizing it.

import pandas as pd
import numpy as np

import sklearn
from sklearn import tree,metrics
from sklearn.neighbors import DistanceMetric
from sklearn.ensemble import RandomForestClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.preprocessing import scale
from sklearn.model_selection import train_test_split
from sklearn.cluster import KMeans, AgglomerativeClustering
from sklearn.neighbors import NearestNeighbors
from sklearn.tree import DecisionTreeClassifier
from sklearn import naive_bayes, svm
from sklearn.naive_bayes import GaussianNB
from sklearn.svm import SVC


from sklearn.model_selection import cross_val_score
from sklearn.model_selection import GridSearchCV
from sklearn import metrics
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, confusion_matrix
from sklearn.metrics import make_scorer

## spliting to train and test

dataset=pd.read_csv("finale_normalized_df.csv")

def split_to_train_and_test(dataset, label_column, test_ratio, rand_state):
    TRAINING_FEATURES = dataset.columns[dataset.columns != label_column]
    TARGET_FEATURE = label_column
    X=dataset[TRAINING_FEATURES]
    y= dataset[TARGET_FEATURE]
    
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=test_ratio, random_state=rand_state)
    return X_train, X_test, y_train, y_test

test_ratio, rand_state = 0.2, 42
category_col_name = 'ratings'
dataset = dataset
X_train, X_test, y_train, y_test = split_to_train_and_test(dataset, category_col_name, test_ratio, rand_state)
print(X_train.shape, X_test.shape, y_train.shape, y_test.shape)

## first Model 
ניסיון ראשון ביצירת מודל, שימוש בשלושה סוגים עם היפרפרמטרים ברירת מחדל ללא אופטימיזציה.

clf1 = tree.DecisionTreeClassifier()
clf2 = RandomForestClassifier()
clf3 = KNeighborsClassifier(n_neighbors=3)

algs = ["DT","RandForest","KNN"]
clfs = [clf1,clf2,clf3]
for i,clf in enumerate(clfs):
    clf.fit(X_train,y_train)
    print(algs[i])
    print("======")
    y_pred = clf.predict(X_test)
    y_pred_train = clf.predict(X_train)
    
    print("accuracy on train data",metrics.accuracy_score(y_true=y_train,y_pred=y_pred_train))
    print("accuracy on test data",metrics.accuracy_score(y_true=y_test,y_pred=y_pred))
    
    print(metrics.confusion_matrix(y_test,y_pred))

הגענו לרמת דיוק של כ-0.78 בק"ננ ועם רמת דיוק על סט המבחן של כ 0.84.
בנוסף ביער ובעץ ההחלטה רואים שיש אוברפיטינג ואחוז דיוק נמוך בהרבה בהשווה לזה של סט הבדיקה
נרצה לנסות וליעל את המודלים.

## Second Model
ננסה למצוא את הפרמטרים הטובים ביותר למודלים במטרה ליעל אותם

dataset=pd.read_csv("finale_normalized_df.csv")

test_ratio, rand_state = 0.2, 42
category_col_name = 'ratings'
dataset = dataset
X_train, X_test, y_train, y_test = split_to_train_and_test(dataset, category_col_name, test_ratio, rand_state)
print(X_train.shape, X_test.shape, y_train.shape, y_test.shape)

ניסינו לבדוק סוגי מודלים נוספים:

clf1 = tree.DecisionTreeClassifier()
clf2 = RandomForestClassifier(random_state=5)
clf3 = KNeighborsClassifier(n_neighbors=5)
clf4 = GaussianNB()
clf5 = SVC()

algs = ["DT","RandForest","KNN","NaiveByse","SVC"]
clfs = [clf1,clf2,clf3,clf4,clf5]
for i,clf in enumerate(clfs):
    clf.fit(X_train,y_train)
    print(algs[i])
    print("======")
    y_pred = clf.predict(X_test)
    y_pred_train = clf.predict(X_train)
    
    print("accuracy on train data",metrics.accuracy_score(y_true=y_train,y_pred=y_pred_train))
    print("accuracy on test data",metrics.accuracy_score(y_true=y_test,y_pred=y_pred))
    
    print(metrics.confusion_matrix(y_test,y_pred))

ניתן לראות כי:

    1. DT: ניתן לראות רמת דיוק של כ-0.95 על סט הטסט אך רק 0.75 על המבחן, הדבר מראה על אוברפיטינג של המודל, ננסה לסדר את הפרמטרים שלו במטרה לקבל מודל תקין.
    
    2.RandForest: ניתן לראות רמת דיוק של כ-0.95 על סט הטסט אך רק 0.78 על המבחן, הדבר מראה על אוברפיטינג של המודל, ננסה לסדר את הפרמטרים שלו במטרה לקבל מודל תקין.
        
    3.KNN: כ-0.78 על סט המבחןת ננסה לעלות את רמת הדיוק על ידי מציאת הפרמטר הטוב ביותר
        
    4.NaiveByse: רמת דיוק של כ-0.3 גם בסט המבחן וגם בסט הטסט, מודל זה אינו מתאים לנו
        
    5.SVC: רמת דיוק של כ 0.80 

#### KNN

ניסינו לשפר את מודל הקנ"נ ואכן הצלחנו למצוא פרמטרים ששיפרו את רמת הדיוק שלו לכ- 0.80

def find_best_k_for_KNN(X_train, y_train):
    parameters= {'n_neighbors': [3,5,7,9,11,15,21,23,19,25,27,29]}
    knn= KNeighborsClassifier()
    clf= GridSearchCV(knn,parameters, scoring=make_scorer(metrics.accuracy_score, greater_is_better=True))
    clf.fit(X_train, y_train)
    best_K= clf.best_params_['n_neighbors']
    best_f1_val= clf.best_score_
    return best_K, best_f1_val

best_K, best_accuracy_score=find_best_k_for_KNN(X_train,y_train)

print("bestk", best_K)

clf3 = KNeighborsClassifier(n_neighbors=29)
clf3.fit(X_train,y_train)
print("KNN")
print("======")
y_pred = clf3.predict(X_test)
y_pred_train = clf3.predict(X_train)
    
print("accuracy on train data",metrics.accuracy_score(y_true=y_train,y_pred=y_pred_train))
print("accuracy on test data",metrics.accuracy_score(y_true=y_test,y_pred=y_pred))
    
print(metrics.confusion_matrix(y_test,y_pred))

#### RandomForst

נתמקד בניסיון לשפר את מודל היער על ידי שינוי פרמטרים:

grid_param ={ 
    "n_estimators":[90,100,115,130],
    "max_depth":range(10,20,1),
    "min_samples_split":range(10,20,1),
    "random_state": range(1,5,1)
}

grid_search = GridSearchCV(estimator=clf2 , param_grid = grid_param, cv=5, n_jobs=-1, verbose=3,scoring=make_scorer(metrics.accuracy_score, greater_is_better=True))

grid_search.fit(X_train, y_train)

grid_search.best_params_

clf = RandomForestClassifier(n_estimators=130, max_depth=15,min_samples_split=15, random_state=2)
clf.fit(X_train,y_train)
print("RandForest")
print("======")
y_pred = clf.predict(X_test)
y_pred_train = clf.predict(X_train)
    
print("accuracy on train data",metrics.accuracy_score(y_true=y_train,y_pred=y_pred_train))
print("accuracy on test data",metrics.accuracy_score(y_true=y_test,y_pred=y_pred))
    
print(metrics.confusion_matrix(y_test,y_pred))

ראינו שיפור ברמת דיור המודל על סט המבחן וירידה בדיוק על סט הבדיקהת מכך ניתן ללמוד כי אנו לא נמצאים במצב של אוברפיטינג יותר.
רמת דיוק על סט המבחן הינה 0.81.

#### DecisionTree

ננסה לשפר גם את מודל העץ החלטה:

grid_param ={
    "min_samples_leaf" : range(1,15,1),
    "max_depth":[5,10,15,30],
    "min_samples_split":range(1,15,1),
}

grid_search = GridSearchCV(estimator=clf1 , param_grid = grid_param, cv=5, n_jobs=-1, verbose=3, scoring=make_scorer(metrics.accuracy_score, greater_is_better=True))

grid_search.fit(X_train, y_train)

grid_search.best_params_

clf1 = DecisionTreeClassifier(max_depth=5,min_samples_split=2,min_samples_leaf= 14 )
clf1.fit(X_train,y_train)
print("DecisionTree")
print("======")
y_pred = clf1.predict(X_test)
y_pred_train = clf1.predict(X_train)
    
print("accuracy on train data",metrics.accuracy_score(y_true=y_train,y_pred=y_pred_train))
print("accuracy on test data",metrics.accuracy_score(y_true=y_test,y_pred=y_pred))
    
print(metrics.confusion_matrix(y_test,y_pred))

#### SVC

param_grid = {'C': [0.1, 1, 10, 100, 1000], 
              'gamma': [1, 0.1, 0.01, 0.001, 0.0001],
              'kernel': ['rbf']} 
  
grid = GridSearchCV(SVC(), param_grid, refit = True, verbose = 3)
  
# fitting the model for grid search
grid.fit(X_train, y_train)

grid.best_params_

clf5 = SVC(C= 10, gamma=0.01 ,kernel='rbf')
clf5.fit(X_train,y_train)
print("SVC")
print("======")
y_pred = clf5.predict(X_test)
y_pred_train = clf5.predict(X_train)
    
print("accuracy on train data",metrics.accuracy_score(y_true=y_train,y_pred=y_pred_train))
print("accuracy on test data",metrics.accuracy_score(y_true=y_test,y_pred=y_pred))
    
print(metrics.confusion_matrix(y_test,y_pred))

לא נראה שיפור משמעותי (0.001)

### F1_score

The F1 score is a popular performance measure for classification and often preferred over, for example,
accuracy when data is unbalanced, such as when the quantity of examples belonging to one class 
significantly outnumbers those found in the other class.

https://c3.ai/glossary/data-science/f1-score/

ניתן לראות שהדאטה שלנו אכן לא מאוזן עם כמות גדולה משמעותית של דוגמאות מדורגות 0 לכן נכון יותר לבדוק את:

#RandomForest
clf = RandomForestClassifier(n_estimators=130, max_depth=15,min_samples_split=15, random_state=2)
clf.fit(X_train,y_train)
y_pred = clf.predict(X_test)
y_pred_train = clf.predict(X_train)

rf_test=f1_score(y_test, y_pred, pos_label='positive',average='weighted')

rf_train=f1_score(y_train, y_pred_train, pos_label='positive',average='weighted')

#svc
clf5.fit(X_train,y_train)
y_pred = clf5.predict(X_test)
y_pred_train = clf5.predict(X_train)

svc_testsvc_train=f1_score(y_test, y_pred, pos_label='positive',average='weighted')

svc_train=f1_score(y_train, y_pred_train, pos_label='positive',average='weighted')

#knn
clf3.fit(X_train,y_train)
y_pred = clf3.predict(X_test)
y_pred_train = clf3.predict(X_train)

knn_test=f1_score(y_test, y_pred, pos_label='positive',average='weighted')

knn_train=f1_score(y_train, y_pred_train, pos_label='positive',average='weighted')

#dt
clf1.fit(X_train,y_train)
y_pred = clf1.predict(X_test)
y_pred_train = clf1.predict(X_train)

dt_test=f1_score(y_test, y_pred, pos_label='positive',average='weighted')

dt_train=f1_score(y_train, y_pred_train, pos_label='positive',average='weighted')

model_name=["RandomForest", "SVC", "KNN", "DecisionTree"]
f1_score_train=[rf_train,svc_train,knn_train,dt_train]
f1_score_test=[rf_test,svc_test,knn_test,dt_test]

df_score=pd.DataFrame({"model_name":model_name , "f1_score train":f1_score_train ,"f1_score test":f1_score_test })
df_score

ניתן לראות כי הציון הכי טוב על גם סט הבדיקה וגם סט המבחן היה במודל היער.

## third Model
ננסה ליעל את המודל על ידי מניפוליזציה של טבלת הנתונים- ננסה להוסיף עמודות היכולות להשפיע על רמת הדיוק וכן ננסה להוריד עמודות בעלות קורליזציה גבוהה אחת לשנייה או בעלות השפעה נמוכה על המודל

ניסינו להוריד את העמודות בעלי ההשפעה הנמוכה ביותר על המודל אך ראינו שבהורדתן אנחנו מורידים מיכולת הדיוק של המודל ולכן בחרנו שלא להוריד אותם.

fDF= pd.DataFrame({"features":clf.feature_names_in_, "weights": clf.feature_importances_ })
fDF.sort_values("weights", ascending=True).head()

tr_df=dataset.copy(deep=True)

temp= dataset["satin"]
tr_df.drop(["satin"], inplace=True, axis=1)

test_ratio, rand_state = 0.2, 42
category_col_name = 'ratings'
dataset = tr_df
X_train, X_test, y_train, y_test = split_to_train_and_test(dataset, category_col_name, test_ratio, rand_state)
clf.fit(X_train,y_train)
y_pred = clf.predict(X_test)
y_pred_train = clf.predict(X_train)
f1_s=f1_score(y_test, y_pred, pos_label='positive',average='weighted')

print("RandomForest")
print("======")
print("accuracy on train data",metrics.accuracy_score(y_true=y_train,y_pred=y_pred_train))
print("accuracy on test data",metrics.accuracy_score(y_true=y_test,y_pred=y_pred))
print("f1_score Score for test: ",f1_s)

print(metrics.confusion_matrix(y_test,y_pred))

tr_df["satin"]=temp

ניסינו להוסיף עמודות הסופרות את מספר המידות ומספר הצבעים הקיימים לאותו הפריט, ההיגיון מאוחרי זה הינו שככל שקיימות יותר אופציות מידה\צבע קהל הקונים של אותו הפריט רחב יותר ועל כן גם מספר הלקוחות עם פוטנציל לדירוג הפריט.

total_size_options=[]
for ind in dataset.index:
    count=0
    if dataset['XXS'][ind]==1: count=count+1
    if dataset['XS'][ind]==1: count=count+1
    if dataset['S'][ind]==1: count=count+1
    if dataset['P'][ind]==1: count=count+1
    if dataset['M'][ind]==1: count=count+1
    if dataset['L'][ind]==1: count=count+1
    if dataset['XL'][ind]==1: count=count+1
    if dataset['XXL'][ind]==1: count=count+1
    if dataset['XXXL'][ind]==1: count=count+1
    if dataset['XXXL'][ind]==1: count=count+1
    if dataset['6XL'][ind]==1: count=count+1
    if dataset['5XL'][ind]==1: count=count+1
    if dataset['One_size'][ind]==1: count=count+1
        
    total_size_options.append(count)
    
dataset["total_size_options"]=total_size_options

test_ratio, rand_state = 0.2, 42
category_col_name = 'ratings'
dataset = dataset
X_train, X_test, y_train, y_test = split_to_train_and_test(dataset, category_col_name, test_ratio, rand_state)
print(X_train.shape, X_test.shape, y_train.shape, y_test.shape)

clf.fit(X_train,y_train)
print("RandomForest")
print("======")
y_pred = clf.predict(X_test)
y_pred_train = clf.predict(X_train)

print("accuracy on train data",metrics.accuracy_score(y_true=y_train,y_pred=y_pred_train))
print("accuracy on test data",metrics.accuracy_score(y_true=y_test,y_pred=y_pred))

print(metrics.confusion_matrix(y_test,y_pred))

dataset.drop(["total_size_options"], inplace=True, axis=1)

ניתן לראות ירידה בדיוק המודל לכן החלטנו להוריד את העמודה הזו.

total_color_options=[]
for ind in dataset.index:
    count=0
    if dataset['Blue'][ind]==1: count=count+1
    if dataset['Black'][ind]==1: count=count+1
    if dataset['Grey'][ind]==1: count=count+1
    if dataset['White'][ind]==1: count=count+1
    if dataset['Brown'][ind]==1: count=count+1
    if dataset['Orange'][ind]==1: count=count+1
    if dataset['Pink'][ind]==1: count=count+1
    if dataset['Green'][ind]==1: count=count+1
    if dataset['Purple'][ind]==1: count=count+1
    if dataset['Red'][ind]==1: count=count+1
    if dataset['Yellow'][ind]==1: count=count+1
    if dataset['other_colors'][ind]==1: count=count+1
    if dataset['One_size'][ind]==1: count=count+1
        
    total_color_options.append(count)

dataset["total_color_options"]=total_color_options

test_ratio, rand_state = 0.2, 42
category_col_name = 'ratings'
dataset = dataset
X_train, X_test, y_train, y_test = split_to_train_and_test(dataset, category_col_name, test_ratio, rand_state)
print(X_train.shape, X_test.shape, y_train.shape, y_test.shape)

clf.fit(X_train,y_train)
print("RandomForest")
print("======")
y_pred = clf.predict(X_test)
y_pred_train = clf.predict(X_train)

print("accuracy on train data",metrics.accuracy_score(y_true=y_train,y_pred=y_pred_train))
print("accuracy on test data",metrics.accuracy_score(y_true=y_test,y_pred=y_pred))

print(metrics.confusion_matrix(y_test,y_pred))

dataset.drop(["total_color_options"], inplace=True, axis=1)

ניתן לראות ירידה בדיוק המודל לכן החלטנו להוריד את העמודה הזו.

ראינו קורלציה גבוהה בין עמודות ה: (אלסטיין) ועמודת ה (פוליאסטר) לכן החלטנו לנסות להוריד אחת מהן ולראות מה ההשפעה על המודל אך גם כאן הייתה ירידה ברמת דיוק המודל

tr_df=dataset.copy(deep=True)
temp= dataset["elastane"]
tr_df.drop(["elastane"], inplace=True, axis=1)

test_ratio, rand_state = 0.2, 42
category_col_name = 'ratings'
dataset = tr_df
X_train, X_test, y_train, y_test = split_to_train_and_test(dataset, category_col_name, test_ratio, rand_state)
clf.fit(X_train,y_train)
print("RandomForest")
print("======")
y_pred = clf.predict(X_test)
y_pred_train = clf.predict(X_train)

print("accuracy on train data",metrics.accuracy_score(y_true=y_train,y_pred=y_pred_train))
print("accuracy on test data",metrics.accuracy_score(y_true=y_test,y_pred=y_pred))

print(metrics.confusion_matrix(y_test,y_pred))

tr_df["elastane"]=temp

קיימת ירידה ברמת הדיוק של סט הבדיקה יחד עם סט המבחן לכן הורדנו גם את העמודה הזו

מהסתכלות בשלב הויזואליזציה, ראינו קורלציה גבוהה עד כמעט זהה בין עמודות

(XL, XXL) ועמודות (XXXL, XXXXL)

לכן בחרנו להוריד את העמודות

(xl, xxxl)

temp1= dataset["XL"]
temp2= dataset["XXXL"]
tr_df.drop(["XL","XXXL"], inplace=True, axis=1)
test_ratio, rand_state = 0.2, 42
category_col_name = 'ratings'
dataset = tr_df
X_train, X_test, y_train, y_test = split_to_train_and_test(dataset, category_col_name, test_ratio, rand_state)
clf.fit(X_train,y_train)
print("RandomForest")
print("======")
y_pred = clf.predict(X_test)
y_pred_train = clf.predict(X_train)

print("accuracy on train data",metrics.accuracy_score(y_true=y_train,y_pred=y_pred_train))
print("accuracy on test data",metrics.accuracy_score(y_true=y_test,y_pred=y_pred))

print(metrics.confusion_matrix(y_test,y_pred))

tr_df["XL"]=temp1
tr_df["XXXL"]=temp2

ראינו שלאחר הורדת עמודות אלו לא ניראה שינוי בדיוק במודל, אך מהסתכלות בפיוזן מטריקס ראינו שיפור בזיהוי 5 כוכבים וירידה בזיהוי 4 כוכבים ולכן החלטנו להשאיר את הטבלה כמו שהיא.

מסקנות:

    הגענו למודל הכי טוב ברמת דיוק של כ- 81% ומכך אנו מסיקות כי ניתן לחזות את דירוג הפריט ברמת דיוק מסוימת.
    
    ניתן לראות מהכונפיוזין מטריקס שהרבה טעויות נעשו בסיווג פריטים אשר שיכים לקטגוריית האפס דבר אשר הגיוני כייון שפריטים רבים מדורגים 0 ישר בכניסתם לאתר ורק בהמשך עולים בדירוג.
    
    המודל נבדק ללא פריטים המדורגים 1 או 2 מה שגם יכול היה להשפיע על רמת הדיוק שלו.
    
    בנוסף ניתן לראות שחלק גדול מ4 הכוכבים מסווגים כ5 או 3 כוכבים וכך גם בדירוד 5 הכוכבים מה שמראה על בעייה ביכולת הבסיווג
    

קישורים לאתרים בהם נעשה שימוש:
https://medium.com/analytics-vidhya/random-forest-classifier-and-its-hyperparameters-8467bec755f6
